Add files via upload
This is  the analyis of the beer dataset in order to find out the  key factors that contribute to overall review in terms of review palette,review taste, review aroma and review appearence. Apart from that sentiment of polarity is used to determine if the review is favorable or not  with the help of TextBlob. EDA is conducted on data set in order to unearth insights that can be beneficial for future business.